# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "XN360 API",
    "version": "1.0.0",
    "description": "OpenAPI spec for Dispatch Queue Items"
  },
  "servers": [
    { "url": "http://localhost:3006/api/v2" }
  ],
  "paths": {
    "/dispatch-queue-items": {
      "get": {
        "summary": "List dispatch queue items (paginated)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "Page number" },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Items per page" },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "default": "created_at" }, "description": "Sort field" },
          { "name": "sort_order", "in": "query", "schema": { "type": "string", "default": "desc", "enum": ["asc", "desc"] }, "description": "Sort order" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of dispatch queue items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "type": { "type": "string" },
                          "payload": { "type": "object" },
                          "status": { "type": "string" },
                          "created_at": { "type": "string", "format": "date-time" },
                          "updated_at": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "current_page": { "type": "integer" },
                        "per_page": { "type": "integer" },
                        "total": { "type": "integer" },
                        "last_page": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a dispatch queue item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string" },
                  "payload": { "type": "object" },
                  "status": { "type": "string" }
                },
                "required": ["type", "payload"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/dispatch-queue-items/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "summary": "Get a dispatch queue item by id",
        "responses": {
          "200": {
            "description": "Single dispatch queue item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "type": { "type": "string" },
                    "payload": { "type": "object" },
                    "status": { "type": "string" },
                    "created_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Update a dispatch queue item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string" },
                  "payload": { "type": "object" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Delete a dispatch queue item",
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DispatchQueueItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "payload": { "type": "object" },
          "status": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}